Major refactor - deleted cruft and standardized package format
