--- choose_your_own/your_algo_randomforest.py
@@ -1,52 +1,0 @@
-#!/usr/bin/python
-
-import matplotlib.pyplot as plt
-from prep_terrain_data import makeTerrainData
-from class_vis import prettyPicture
-from time import time
-
-features_train, labels_train, features_test, labels_test = makeTerrainData()
-
-
-### the training data (features_train, labels_train) have both "fast" and "slow"
-### points mixed together--separate them so we can give them different colors
-### in the scatterplot and identify them visually
-grade_fast = [features_train[ii][0] for ii in range(0, len(features_train)) if labels_train[ii]==0]
-bumpy_fast = [features_train[ii][1] for ii in range(0, len(features_train)) if labels_train[ii]==0]
-grade_slow = [features_train[ii][0] for ii in range(0, len(features_train)) if labels_train[ii]==1]
-bumpy_slow = [features_train[ii][1] for ii in range(0, len(features_train)) if labels_train[ii]==1]
-
-
-#### initial visualization
-plt.xlim(0.0, 1.0)
-plt.ylim(0.0, 1.0)
-plt.scatter(bumpy_fast, grade_fast, color = "b", label="fast")
-plt.scatter(grade_slow, bumpy_slow, color = "r", label="slow")
-plt.legend()
-plt.xlabel("bumpiness")
-plt.ylabel("grade")
-# plt.show()
-################################################################################
-
-
-### your code here!  name your classifier object clf if you want the
-### visualization code (prettyPicture) to show you the decision boundary
-from sklearn import ensemble
-n_estimators = 100
-clf = ensemble.RandomForestClassifier(n_estimators=n_estimators)
-print "\nn_estimators =", n_estimators
-
-t0 = time()
-clf.fit(features_train, labels_train)
-# print "training time:", round(time()-t0, 3), "s"
-
-t1 = time()
-print "SCORE:", clf.score(features_test, labels_test)
-# print "scoring time:", round(time()-t1, 3), "s"
-
-
-try:
-    prettyPicture(clf, features_test, labels_test)
-except NameError as err:
-    print "ERROR:", err
-    pass
