--- final_project/common_configs.py
@@ -1,8 +1,0 @@
-VERBOSE = False
-
-DATASET_DICTIONARY_FILE = "resources/final_project_dataset.pkl"
-
-EXPORT_LOG_FILENAME = "output/log.txt"
-EXPORT_CLF_FILENAME = "output/model.pkl"
-EXPORT_DATASET_FILENAME = "output/dataset.pkl"
-EXPORT_FEATURE_LIST_FILENAME = "output/feature_list.txt"
