--- final_project/train_and_export_to_gcp.py
@@ -1,13 +1,0 @@
-#!/usr/bin/python
-
-import poi_id
-import export_to_gcp
-
-
-def main():
-    poi_id.main()
-    export_to_gcp.main()
-
-
-if __name__ == '__main__':
-    main()
