--- outliers/practice_outliers_ages.pkl
@@ -1,102 +1,0 @@
-(lp0
-I49
-aI21
-aI32
-aI30
-aI53
-aI51
-aI61
-aI23
-aI39
-aI21
-aI30
-aI43
-aI21
-aI29
-aI49
-aI45
-aI30
-aI47
-aI57
-aI20
-aI57
-aI52
-aI35
-aI27
-aI64
-aI35
-aI24
-aI24
-aI58
-aI47
-aI57
-aI53
-aI44
-aI64
-aI37
-aI45
-aI58
-aI48
-aI59
-aI46
-aI52
-aI22
-aI30
-aI33
-aI23
-aI30
-aI24
-aI32
-aI49
-aI36
-aI37
-aI29
-aI32
-aI63
-aI49
-aI48
-aI27
-aI53
-aI27
-aI37
-aI65
-aI49
-aI45
-aI51
-aI58
-aI55
-aI30
-aI21
-aI34
-aI32
-aI29
-aI63
-aI60
-aI34
-aI50
-aI38
-aI62
-aI41
-aI32
-aI31
-aI45
-aI32
-aI46
-aI61
-aI38
-aI30
-aI65
-aI43
-aI24
-aI22
-aI25
-aI48
-aI56
-aI39
-aI22
-aI37
-aI65
-aI44
-aI64
-aI59
-a.
