--- outliers/practice_outliers_net_worths.pkl
@@ -1,102 +1,0 @@
-(lp0
-F316.18428306022463
-aF128.4847139765763
-aF212.95377076201385
-aF217.96059712816052
-aF326.56693250553326
-aF314.0672608610164
-aF412.83425631014785
-aF159.09869458305818
-aF234.36051228130097
-aF142.1012008717193
-aF178.23164614375077
-aF259.43540492859483
-aF136.08924543132068
-aF142.98439510684403
-aF271.75164334973937
-aF270.00424941518054
-aF167.24337759331152
-aF300.0349466519055
-aF338.08951848957577
-aF96.75392597329417
-aF385.5629753784311
-aF320.4844739902693
-aF220.10056409375846
-aF140.25503627573087
-aF389.1123455094963
-aF220.96845179419734
-aF126.98012845155395
-aF157.51396036691344
-aF350.4872262016239
-aF287.91612500413447
-aF344.21586775541203
-aF368.29556369017877
-aF274.7300555052413
-aF378.84578142088196
-aF247.70089824206377
-aF256.83312700057957
-aF366.6772719000951
-aF260.8065975224045
-aF342.1862790220314
-aF291.4372247173825
-aF339.7693315999082
-aF140.9273656237994
-aF185.1870343522352
-aF200.22792608821422
-aF114.17956019265145
-aF173.10311583210583
-aF140.78722458080426
-aF221.1424445243783
-aF313.12236579136925
-aF189.73919689274533
-aF237.7316793878959
-aF173.54835439167368
-aF186.46155999388083
-aF405.9835257768174
-aF326.86999044991904
-aF318.62560238232396
-aF151.96564953554724
-aF325.0657524829757
-aF175.37526862807127
-aF250.76090254244718
-aF396.6665152430942
-aF302.53682046672367
-aF259.1233005198794
-aF294.8258675183866
-aF378.75051644788397
-aF370.8748005714165
-aF186.05979756839332
-aF151.32065795784047
-aF219.7327205009527
-aF187.0976049078975
-aF188.4779121101683
-aF424.51073132931936
-aF374.28347921780096
-aF243.79287311628013
-aF260.1050979182051
-aF253.93805008750448
-aF389.24094136476344
-aF250.26985299068266
-aF201.83521553071006
-aF153.99862170798215
-aF276.85656224324975
-aF207.14225143023492
-aF317.0578808948303
-aF370.88459563452705
-aF221.33012794213624
-aF177.46485912830926
-aF424.5580423540415
-aF275.3250221931937
-aF139.40479592465923
-aF147.76534866226712
-aF133.88384977349668
-aF80.9013362873476
-aF121.11602327460938
-aF134.41369498852399
-aF47.70052124577958
-aF16.507788679151513
-aF34.19027438129125
-aF64.06616829393845
-aF122.70221488837397
-aF129.10958748845152
-a.
