+++ src/trainer/poi_email_addresses.py
@@ -1,0 +1,94 @@
+def poiEmails():
+    email_list = ["kenneth_lay@enron.net",
+                  "kenneth_lay@enron.com",
+                  "klay.enron@enron.com",
+                  "kenneth.lay@enron.com",
+                  "klay@enron.com",
+                  "layk@enron.com",
+                  "chairman.ken@enron.com",
+                  "jeffreyskilling@yahoo.com",
+                  "jeff_skilling@enron.com",
+                  "jskilling@enron.com",
+                  "effrey.skilling@enron.com",
+                  "skilling@enron.com",
+                  "jeffrey.k.skilling@enron.com",
+                  "jeff.skilling@enron.com",
+                  "kevin_a_howard.enronxgate.enron@enron.net",
+                  "kevin.howard@enron.com",
+                  "kevin.howard@enron.net",
+                  "kevin.howard@gcm.com",
+                  "michael.krautz@enron.com"
+                  "scott.yeager@enron.com",
+                  "syeager@fyi-net.com",
+                  "scott_yeager@enron.net",
+                  "syeager@flash.net",
+                  "joe'.'hirko@enron.com",
+                  "joe.hirko@enron.com",
+                  "rex.shelby@enron.com",
+                  "rex.shelby@enron.nt",
+                  "rex_shelby@enron.net",
+                  "jbrown@enron.com",
+                  "james.brown@enron.com",
+                  "rick.causey@enron.com",
+                  "richard.causey@enron.com",
+                  "rcausey@enron.com",
+                  "calger@enron.com",
+                  "chris.calger@enron.com",
+                  "christopher.calger@enron.com",
+                  "ccalger@enron.com",
+                  "tim_despain.enronxgate.enron@enron.net",
+                  "tim.despain@enron.com",
+                  "kevin_hannon@enron.com",
+                  "kevin'.'hannon@enron.com",
+                  "kevin_hannon@enron.net",
+                  "kevin.hannon@enron.com",
+                  "mkoenig@enron.com",
+                  "mark.koenig@enron.com",
+                  "m..forney@enron.com",
+                  "ken'.'rice@enron.com",
+                  "ken.rice@enron.com",
+                  "ken_rice@enron.com",
+                  "ken_rice@enron.net",
+                  "paula.rieker@enron.com",
+                  "prieker@enron.com",
+                  "andrew.fastow@enron.com",
+                  "lfastow@pdq.net",
+                  "andrew.s.fastow@enron.com",
+                  "lfastow@pop.pdq.net",
+                  "andy.fastow@enron.com",
+                  "david.w.delainey@enron.com",
+                  "delainey.dave@enron.com",
+                  "'delainey@enron.com",
+                  "david.delainey@enron.com",
+                  "'david.delainey'@enron.com",
+                  "dave.delainey@enron.com",
+                  "delainey'.'david@enron.com",
+                  "ben.glisan@enron.com",
+                  "bglisan@enron.com",
+                  "ben_f_glisan@enron.com",
+                  "ben'.'glisan@enron.com",
+                  "jeff.richter@enron.com",
+                  "jrichter@nwlink.com",
+                  "lawrencelawyer@aol.com",
+                  "lawyer'.'larry@enron.com",
+                  "larry_lawyer@enron.com",
+                  "llawyer@enron.com",
+                  "larry.lawyer@enron.com",
+                  "lawrence.lawyer@enron.com",
+                  "tbelden@enron.com",
+                  "tim.belden@enron.com",
+                  "tim_belden@pgn.com",
+                  "tbelden@ect.enron.com",
+                  "michael.kopper@enron.com",
+                  "dave.duncan@enron.com",
+                  "dave.duncan@cipco.org",
+                  "duncan.dave@enron.com",
+                  "ray.bowen@enron.com",
+                  "raymond.bowen@enron.com",
+                  "'bowen@enron.com",
+                  "wes.colwell@enron.com",
+                  "dan.boyle@enron.com",
+                  "cloehr@enron.com",
+                  "chris.loehr@enron.com"
+                  ]
+    return email_list
