--- tools/python2_lesson06_keys.pkl
@@ -1,288 +1,0 @@
-(lp0
-S'METTS MARK'
-p1
-aS'BAXTER JOHN C'
-p2
-aS'ELLIOTT STEVEN'
-p3
-aS'CORDES WILLIAM R'
-p4
-aS'HANNON KEVIN P'
-p5
-aS'MORDAUNT KRISTINA M'
-p6
-aS'MEYER ROCKFORD G'
-p7
-aS'MCMAHON JEFFREY'
-p8
-aS'HORTON STANLEY C'
-p9
-aS'PIPER GREGORY F'
-p10
-aS'HUMPHREY GENE E'
-p11
-aS'UMANOFF ADAM S'
-p12
-aS'BLACHMAN JEREMY M'
-p13
-aS'SUNDE MARTIN'
-p14
-aS'GIBBS DANA R'
-p15
-aS'LOWRY CHARLES P'
-p16
-aS'COLWELL WESLEY'
-p17
-aS'MULLER MARK S'
-p18
-aS'JACKSON CHARLENE R'
-p19
-aS'WESTFAHL RICHARD K'
-p20
-aS'WALTERS GARETH W'
-p21
-aS'WALLS JR ROBERT H'
-p22
-aS'KITCHEN LOUISE'
-p23
-aS'CHAN RONNIE'
-p24
-aS'BELFER ROBERT'
-p25
-aS'SHANKMAN JEFFREY A'
-p26
-aS'WODRASKA JOHN'
-p27
-aS'BERGSIEKER RICHARD P'
-p28
-aS'URQUHART JOHN A'
-p29
-aS'BIBI PHILIPPE A'
-p30
-aS'RIEKER PAULA H'
-p31
-aS'WHALEY DAVID A'
-p32
-aS'BECK SALLY W'
-p33
-aS'HAUG DAVID L'
-p34
-aS'ECHOLS JOHN B'
-p35
-aS'MENDELSOHN JOHN'
-p36
-aS'HICKERSON GARY J'
-p37
-aS'CLINE KENNETH W'
-p38
-aS'LEWIS RICHARD'
-p39
-aS'HAYES ROBERT E'
-p40
-aS'MCCARTY DANNY J'
-p41
-aS'KOPPER MICHAEL J'
-p42
-aS'LEFF DANIEL P'
-p43
-aS'LAVORATO JOHN J'
-p44
-aS'BERBERIAN DAVID'
-p45
-aS'DETMERING TIMOTHY J'
-p46
-aS'WAKEHAM JOHN'
-p47
-aS'POWERS WILLIAM'
-p48
-aS'GOLD JOSEPH'
-p49
-aS'BANNANTINE JAMES M'
-p50
-aS'DUNCAN JOHN H'
-p51
-aS'SHAPIRO RICHARD S'
-p52
-aS'SHERRIFF JOHN R'
-p53
-aS'SHELBY REX'
-p54
-aS'LEMAISTRE CHARLES'
-p55
-aS'DEFFNER JOSEPH M'
-p56
-aS'KISHKILL JOSEPH G'
-p57
-aS'WHALLEY LAWRENCE G'
-p58
-aS'MCCONNELL MICHAEL S'
-p59
-aS'PIRO JIM'
-p60
-aS'DELAINEY DAVID W'
-p61
-aS'SULLIVAN-SHAKLOVITZ COLLEEN'
-p62
-aS'WROBEL BRUCE'
-p63
-aS'LINDHOLM TOD A'
-p64
-aS'MEYER JEROME J'
-p65
-aS'BUTTS ROBERT H'
-p66
-aS'OLSON CINDY K'
-p67
-aS'MCDONALD REBECCA'
-p68
-aS'CUMBERLAND MICHAEL S'
-p69
-aS'GAHN ROBERT S'
-p70
-aS'MCCLELLAN GEORGE'
-p71
-aS'HERMANN ROBERT J'
-p72
-aS'SCRIMSHAW MATTHEW'
-p73
-aS'GATHMANN WILLIAM D'
-p74
-aS'HAEDICKE MARK E'
-p75
-aS'BOWEN JR RAYMOND M'
-p76
-aS'GILLIS JOHN'
-p77
-aS'FASTOW ANDREW S'
-p78
-aS'FITZGERALD JAY L'
-p79
-aS'MORAN MICHAEL P'
-p80
-aS'REDMOND BRIAN L'
-p81
-aS'BAZELIDES PHILIP J'
-p82
-aS'BELDEN TIMOTHY N'
-p83
-aS'DURAN WILLIAM D'
-p84
-aS'THORN TERENCE H'
-p85
-aS'FOY JOE'
-p86
-aS'CALGER CHRISTOPHER F'
-p87
-aS'RICE KENNETH D'
-p88
-aS'KAMINSKI WINCENTY J'
-p89
-aS'LOCKHART EUGENE E'
-p90
-aS'COX DAVID'
-p91
-aS'OVERDYKE JR JERE C'
-p92
-aS'PEREIRA PAULO V. FERRAZ'
-p93
-aS'STABLER FRANK'
-p94
-aS'BLAKE JR. NORMAN P'
-p95
-aS'SHERRICK JEFFREY B'
-p96
-aS'PRENTICE JAMES'
-p97
-aS'GRAY RODNEY'
-p98
-aS'PICKERING MARK R'
-p99
-aS'THE TRAVEL AGENCY IN THE PARK'
-p100
-aS'NOLES JAMES L'
-p101
-aS'KEAN STEVEN J'
-p102
-aS'FOWLER PEGGY'
-p103
-aS'WASAFF GEORGE'
-p104
-aS'WHITE JR THOMAS E'
-p105
-aS'CHRISTODOULOU DIOMEDES'
-p106
-aS'ALLEN PHILLIP K'
-p107
-aS'SHARP VICTORIA T'
-p108
-aS'JAEDICKE ROBERT'
-p109
-aS'WINOKUR JR. HERBERT S'
-p110
-aS'BROWN MICHAEL'
-p111
-aS'BADUM JAMES P'
-p112
-aS'HUGHES JAMES A'
-p113
-aS'REYNOLDS LAWRENCE'
-p114
-aS'DIMICHELE RICHARD G'
-p115
-aS'BHATNAGAR SANJAY'
-p116
-aS'CARTER REBECCA C'
-p117
-aS'BUCHANAN HAROLD G'
-p118
-aS'YEAP SOON'
-p119
-aS'MURRAY JULIA H'
-p120
-aS'GARLAND C KEVIN'
-p121
-aS'DODSON KEITH'
-p122
-aS'YEAGER F SCOTT'
-p123
-aS'HIRKO JOSEPH'
-p124
-aS'DIETRICH JANET R'
-p125
-aS'DERRICK JR. JAMES V'
-p126
-aS'FREVERT MARK A'
-p127
-aS'PAI LOU L'
-p128
-aS'BAY FRANKLIN R'
-p129
-aS'HAYSLETT RODERICK J'
-p130
-aS'FUGH JOHN L'
-p131
-aS'FALLON JAMES B'
-p132
-aS'KOENIG MARK E'
-p133
-aS'SAVAGE FRANK'
-p134
-aS'IZZO LAWRENCE L'
-p135
-aS'TILNEY ELIZABETH A'
-p136
-aS'MARTIN AMANDA K'
-p137
-aS'BUY RICHARD B'
-p138
-aS'GRAMM WENDY L'
-p139
-aS'CAUSEY RICHARD A'
-p140
-aS'TAYLOR MITCHELL S'
-p141
-aS'DONAHUE JR JEFFREY M'
-p142
-aS'GLISAN JR BEN F'
-p143
-a.
