+++ train-cmle.sh
@@ -8,8 +8,8 @@ source "/home/nkarasch/anaconda3/bin/activate" udacity
 PROJECT_ID="adept-elf-206419"
 BUCKET_ID="${PROJECT_ID}-mlengine"
 REGION="us-central1"
-TRAINING_PACKAGE_PATH="/home/nkarasch/GitStuff/ud120-projects/final_project/"
-MAIN_TRAINER_MODULE="final_project.train_and_export_to_gcp"
+TRAINING_PACKAGE_PATH="/home/nkarasch/GitStuff/ud120-projects/src/trainer"
+MAIN_TRAINER_MODULE="src.trainer.train_and_export_to_gcp"
 MODEL_NAME="enron_poi_classifier"
 
 gcloud ml-engine local train \
